Fix src/content compilation errors for 64-bit mac builds
TBR=mac64

Review URL: https://chromiumcodereview.appspot.com/11320002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@164668 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f3fde867a6ad1210f1ad3ded55d82810b9db065a

+++ content/browser/renderer_host/accelerated_plugin_view_mac.mm
@@ -83,7 +83,7 @@ using content::RenderWidgetHostViewMac;
 - (void)drawRect:(NSRect)rect {
   TRACE_EVENT0("browser", "AcceleratedPluginView::drawRect");
   const NSRect* dirtyRects;
-  int dirtyRectCount;
+  NSInteger dirtyRectCount;
   [self getRectsBeingDrawn:&dirtyRects count:&dirtyRectCount];
 
   {
